% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_func.R
\name{lm_func}
\alias{lm_func}
\title{linear regression function}
\usage{
lm_func(X, y, weights, na.action = c("ignore", "mean_impute", "mice_impute"))
}
\arguments{
\item{X}{A data frame (one or multiple columns) of predictors (both continuous and categorical)}

\item{y}{A vector of outcome values (continuous)}

\item{na.action}{Possible treatment for NA values, including "ignore", "mean_impute", "mice_impute". Default to "ignore"}
}
\value{
A list of useful attributes calculated from the linear model
}
\description{
Calculate linear model results including coefficient, standard errors, t and p values, MSE, Rsquared and Adusted Rsquared
}
\examples{
require("MASS")
data(Boston)
X = Boston[-ncol(Boston)]
y = Boston$medv
lm_func(X,y, na.action = "ignore")
}
