% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_func.R
\name{qq_plot}
\alias{qq_plot}
\title{QQ Plot of the Standardized Residuals}
\usage{
qq_plot(lm_object)
}
\arguments{
\item{lm_object}{A linear models object obtained from the linear model function "lm_func" in the package}
}
\value{
A QQ Plot of the Standardized Residuals
}
\description{
Create the QQ Plot of the Standardized Residuals; similar to the diagnostic plots of the default linear models
}
\examples{
require("MASS")
data(Boston)
X = Boston[-ncol(Boston)]
y = Boston$medv
lm_result = lm_func(X,y, na.action = "ignore")
qq_plot(lm_object)
}
