% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_func.R
\name{res_lev}
\alias{res_lev}
\title{Residual VS Leverage Graph}
\usage{
res_lev(lm_object)
}
\arguments{
\item{lm_object}{A linear models object obtained from the linear model function "lm_func" in the package}
}
\value{
A Residual VS Leverage plot
}
\description{
Create the Residual VS Leverage plot; similar to the diagnostic plots of the default linear models
}
\examples{
require("MASS")
data(Boston)
X = Boston[-ncol(Boston)]
y = Boston$medv
lm_result = lm_func(X,y, na.action = "ignore")
res_lev(lm_object)
}
