% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_func.R
\name{scale_location}
\alias{scale_location}
\title{Scale-Location Graph}
\usage{
scale_location(lm_object)
}
\arguments{
\item{lm_object}{A linear models object obtained from the linear model function "lm_func" in the package}
}
\value{
A Scale-Location plot
}
\description{
Create the Scale-Location plot; similar to the diagnostic plots of the default linear models
}
\examples{
require("MASS")
data(Boston)
X = Boston[-ncol(Boston)]
y = Boston$medv
lm_result = lm_func(X,y, na.action = "ignore")
scale_location(lm_object)
}
