% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treat_na.R
\name{treat_na}
\alias{treat_na}
\title{Missing value imputation}
\usage{
treat_na(na.action, data, y)
}
\arguments{
\item{na.action}{Possible treatment for NA values, including "ignore", "mean_impute", "mice_impute". Default to "ignore"}

\item{data}{A data frame (one or multiple columns) of predictors (continous or categorical) with or without missing values}

\item{y}{A vector of outcome values (continuous) with or without missing values}
}
\value{
A list of data without missing values that can be used in the linear model; objects include complete predictor matrix and complete outcome variable
}
\description{
Function to treat missing values; also embedded in the lm_func
}
\examples{
n = 10000
x1 = rnorm(n, mean = 100, sd = 2)
e = rnorm(n, mean = 0, sd = 1)
y = 5 + 3 * x1 + e
## Randomly generate some missing values
set.seed(123)
ind = sample(x = n , size = 0.1 * n)
x1[ind] = NA
## Creating a data frame for storing testing data
X = data.frame(x1)
treat_na(na.action = "mice_impute",X,y)
}
